Split events to their respective folders, adjust comments.
