+++ src/events/account.ts
@@ -2,8 +2,8 @@
 // Events which are only usable while logged in.
 // ******************** //
 
-import fetchProfileId from './account/fetchProfileId';
-import fetchProfileData from './account/fetchProfileData';
-import logout from './account/logout';
+import fetchProfileId from './account/fetchProfileId/fetchProfileId';
+import fetchProfileData from './account/fetchProfileData/fetchProfileData';
+import logout from './account/logout/logout';
 
 export default { fetchProfileId, fetchProfileData, logout }
