+++ src/events/account/fetchProfileData/interfaces.ts
@@ -1,10 +1,8 @@
 // ******************** //
-// Interfaces for the account-only event files.
+// Interfaces for the fetchProfileData account-only event file.
 // ******************** //
 
 import { Account, EventArguments } from 'other/interfaces';
-
-// TODO: each event will have its own folder with interfaces and the event file.
 
 // fetchProfileData
 export interface FetchProfileData extends EventArguments {
@@ -15,7 +13,4 @@ export interface FetchProfileDataResult {
     profileData: Partial<Account>;
 }
 
-// fetchProfileId
-export interface FetchProfileIdResult  {
-    profileId: string;
-}
+
