+++ src/events/account/fetchProfileId/interfaces.ts
@@ -1,0 +1,7 @@
+// ******************** //
+// Interfaces for the fetchProfileId account-only event file.
+// ******************** //
+
+export interface FetchProfileIdResult  {
+    profileId: string;
+}
