+++ src/events/account/logout/logout.ts
