+++ src/events/general.ts
@@ -2,6 +2,6 @@
 // Events which are usable regardless of login state.
 // ******************** //
 
-import isLoggedIn from './general/isLoggedIn';
+import isLoggedIn from './general/isLoggedIn/isLoggedIn';
 
 export default { isLoggedIn }
