+++ src/events/general/isLoggedIn/isLoggedIn.ts
@@ -4,7 +4,7 @@
 
 import { EventArguments } from 'other/interfaces';
 import { isSocketLoggedIn } from 'other/utilities';
-import { IsLoggedInResult } from './interface';
+import { IsLoggedInResult } from './interfaces';
 
 export default function isLoggedIn({ socket }: EventArguments): IsLoggedInResult {
     return {loggedIn: isSocketLoggedIn(socket)};
