+++ src/events/noAccount.ts
@@ -2,8 +2,8 @@
 // Events which are only usable while not logged in.
 // ******************** //
 
-import register from './noAccount/register';
-import login from './noAccount/login';
-import loginToken from './noAccount/loginToken';
+import register from './noAccount/register/register';
+import login from './noAccount/login/login';
+import loginToken from './noAccount/loginToken/loginToken';
 
 export default { register, login, loginToken }
